requirements.txt
kivy==2.3.0
pyjnius==1.6.1
buildozer==1.5.0
